requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.0.6
asgiref==3.7.2
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
authclient==1.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2019.11.28
cffi==1.15.1
channels==3.0.4
channels-redis==4.1.0
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
distro==1.4.0
Django==3.2.7
django-3-jet==1.0.8
django-background-task==0.1.8
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.10.1
django-filter==21.1
django-import-export==2.7.1
django-jet==1.0.8
django-js-asset==1.2.2
django-timezone-field==4.2.1
djangochannelsrestframework==1.0.0
djangorestframework==3.14.0
drf-yasg==1.21.5
ecdsa==0.17.0
entrypoints==0.3
enum-compat==0.0.3
et-xmlfile==1.1.0
Flask==2.0.1
frozenlist==1.2.0
future==0.18.2
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
httplib2==0.14.0
hyperlink==21.0.0
idna==2.8
idna-ssl==1.1.0
importlib-metadata==4.8.3
incremental==22.10.0
inflection==0.5.1
intuit-oauth==1.2.4
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
keyring==18.0.1
kombu==5.1.0
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupPy==1.14
MarkupSafe==2.0.1
msgpack==1.0.7
multidict==5.2.0
mysqlclient==2.1.1
numpy==1.19.5
oauthlib==3.1.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.0
pandas==1.1.5
passlib==1.7.4
pbr==5.11.1
Pillow==7.0.0
prompt-toolkit==3.0.20
protobuf==3.6.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.3.0
pycodestyle==2.7.0
pycparser==2.18
pycryptodome==3.10.4
Pygments==2.9.0
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==23.2.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.7.3
python-jose==3.3.0
python-quickbooks==0.9.0
pytz==2023.3
pyxdg==0.26
PyYAML==5.3.1
rauth==0.7.3
razorpay==1.3.0
redis==5.0.1
reportlab==3.5.34
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
SecretStorage==2.3.1
service-identity==23.1.0
simplejson==3.17.6
six==1.14.0
sqlparse==0.4.1
tablib==3.1.0
testresources==2.0.1
toml==0.10.2
twisted==23.8.0
txaio==23.1.1
typing-extensions==4.0.0
uritemplate==3.0.1
urllib3==1.25.8
vine==5.0.0
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==2.0.1
WTForms==2.3.3
xlrd==2.0.1
xlwt==1.3.0
yarl==1.7.2
zipp==3.6.0
zope.interface==6.1
